GridDataOutput: add translation unit
