+++ CMakeLists_files.cmake
@@ -621,8 +621,10 @@ if (Damaris_FOUND AND MPI_FOUND)
   list (APPEND PUBLIC_HEADER_FILES ebos/damaris_properties.hh)
   list (APPEND PUBLIC_HEADER_FILES ebos/damariswriter.hh)
   list (APPEND PUBLIC_HEADER_FILES opm/simulators/utils/DamarisVar.hpp)
-  list (APPEND PUBLIC_HEADER_FILES opm/simulators/utils/GridDataOutput.hpp)
-  list(APPEND MAIN_SOURCE_FILES opm/simulators/utils/DamarisVar.cpp)
+  list (APPEND PUBLIC_HEADER_FILES opm/simulators/utils/GridDataOutput.hpp
+                                   opm/simulators/utils/GridDataOutput_impl.hpp)
+  list(APPEND MAIN_SOURCE_FILES opm/simulators/utils/DamarisVar.cpp
+                                opm/simulators/utils/GridDataOutput.cpp)
 endif()
 
 if(HDF5_FOUND)
