+++ flow/flow_blackoil_polyhedralgrid.cpp
@@ -33,6 +33,7 @@
 #include <ebos/ecltransmissibility_impl.hh>
 #include <ebos/eclgenericwriter_impl.hh>
 #include <ebos/equil/initstateequil_impl.hh>
+#include <opm/simulators/utils/GridDataOutput_impl.hpp>
 
 namespace Opm {
 namespace Properties {
